refactor: (#110) Dao 반환 값을 래퍼 클래스가 아닌 원시형으로 변경한다.

+++ src/main/java/spring/backend/activity/dto/response/ActivityWithTitleAndSavedTimeResponse.java
@@ -9,7 +9,7 @@ public record ActivityWithTitleAndSavedTimeResponse(
         String title,
 
         @Schema(description = "모은 시간", example = "60")
-        int savedTime,
+        long savedTime,
 
         @Schema(description = "활동 날짜", example = "2021-07-01T00:00:00")
         LocalDateTime dateOfActivity
